Move types and delete dead code
